[UNTESTED] Separate gradient board class
Experiment class returns status too
